rename cards.jsx to hand.jsx
